refactor(astronomy): lift the shared ephemeris out of EclipseCalculator
EclipseCalculator held a complete Meeus lunar and solar theory — chapter 47's
truncated ELP-2000/82 series, topocentric parallax, nutation, sidereal time and
ΔT — and every bit of it was private to that file. None of it is specific to
eclipses, and none of it could be used for anything else.

Move it to Ephemeris. EclipseCalculator keeps only the eclipse geometry:
finding lunations whose shadow reaches the Earth, and what an observer at one
place sees of it.

The move is numerically inert. Counting every numeric literal across both files
gives 937 before and 937 after; the only differences are that the lunation index
went from being spelled out twice to once, and two new constants for the mean
lunar radius and the synodic month. EclipseCalculatorTests is unchanged and is
the real check.

One constant deliberately does not become the default. The moon radius here was
the umbral k = 0.272281, chosen so totality durations match NASA's, and it is
the wrong figure for how large the moon looks. Ephemeris exposes the mean radius
as moonRadiusKm and keeps the eclipse value as umbralMoonRadiusKm, so eclipse
contacts still use it and nothing else picks it up by accident.
